Major bug fixed with decimal and fraction creation
problem_set.make_decimal() did not account for the case where a user had
selected a single value. Now that single value is treated as a lower
bound for the input with a range of 1.

problem_set.make_integer() also did not account for this case. The try
statement may be redundant because the range property will always be
false in this case so different logic would be used. Will leave this in
for cases where problem settings are created manually and erroneously.