Major breakthrough in problem_set switching functionality.
problem_element_manager had an early implementation of disconnect().
The buggy implementation used a null check to avoid raising a TypeError.
This was a false equivalency and has been fixed by adding a try except
block. Should really get around to adding logging into the project so
that the except clause cases are recorded (possibly with a record of the
stack trace leading up to the logged line). The skipped disconnection of
the signals led to the default integer_selection view being loaded
instead of the view that corresponded to the current model. Need to do
further testing to see if problem_set switching is completely
functional. I think we're getting close!