add setting bug fixed. term count change tested.
Fixed a bug where adding a element_setting_group to a term or an
operator resulted in the last selected element_setting_group being
cleared. This fix relies on the add element_setting_group only being
available when a topLevelItem is selected (desired behavior, but if
removed, another fix can be made).

Further tested adding and removing terms. Need to do more edge-case
testing, but everything works so far.